Parser test

+++ Code/E1_PrepareEvalData.py
@@ -70,9 +70,9 @@ if Track=='FEDRA':
  final_rows=len(data.axes[0])
  print(UF.TimeStamp(),'The cleaned data has ',final_rows,' hits')
  data[PM.MC_Event_ID] = data[PM.MC_Event_ID].astype(str)
- #data[PM.FEDRA_Track_ID] = data[PM.FEDRA_Track_ID].astype(int)
+ data[PM.FEDRA_Track_ID] = data[PM.FEDRA_Track_ID].astype(int)
  data[PM.FEDRA_Track_ID] = data[PM.FEDRA_Track_ID].astype(str)
- data[PM.FEDRA_Track_QUADRANT] = data[PM.FEDRA_Track_QUADRANT].astype(int)
+ #data[PM.FEDRA_Track_QUADRANT] = data[PM.FEDRA_Track_QUADRANT].astype(int)
  data[PM.FEDRA_Track_QUADRANT] = data[PM.FEDRA_Track_QUADRANT].astype(str)
  data['Track_ID'] = data[PM.FEDRA_Track_QUADRANT] + '-' + data[PM.FEDRA_Track_ID]
  data['Mother_ID'] = data[PM.MC_Event_ID] + '-' + data[PM.MC_VX_ID]
